Updated List_AllInputs_
I updated the List_AllInputs_MatEdits file with additional points for boiler, rtus, and pumps. I need to have a discussion with someone in the Springfield office before selecting the right VAV boxes to move forward with. Also, added some drawing sets to the folder to help with the conversation about VAV boxes.